Acabodo - mathias

+++ src/Recursividad/ejercicios/Potencia.java
@@ -11,7 +11,7 @@ public class Potencia extends Variables {
         setSuma(1);
     }
 
-    public int potencia(){
+    public long potencia(){
         if(getCiclo() <= Math.abs(getNumero2())){
           setSuma(getSuma() * getNumero());
             setCiclo(getCiclo() +1);
